old/deploy/requirements.txt
google-generativeai
requests
python-dotenv
runpod
